Chinh Migration va xu lu hoa don xuat

+++ app/Http/Controllers/LoaiSanPhamController.php
@@ -5,10 +5,14 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\LoaiSanPham;
 use App\Http\Requests\ThemMoiLoaiSanPhamRequest;
+use Illuminate\Support\Facades\Gate;
 class LoaiSanPhamController extends Controller
 {
     public function ThemMoiLoaiSp()
     {
+        if (Gate::denies('quan-ly-san-pham')) {
+            return redirect()->route('trang-chu')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
         return view('loai-san-pham/them-moi');
     }
     public function XuLyThemMoiLoai(ThemMoiLoaiSanPhamRequest $request)
@@ -28,22 +32,33 @@ class LoaiSanPhamController extends Controller
 
     }
 
-    public function DanhSachLoaiSp(){
+    public function DanhSachLoaiSp()
+    {
         $dsLoaiSp=LoaiSanPham::all();
         return view('loai-san-pham/danh-sach',compact('dsLoaiSp'));
     }
-    public function XoaLoaiSp($id){
+    public function XoaLoaiSp($id)
+    {
+        if (Gate::denies('quan-ly-san-pham')) {
+            return redirect()->route('trang-chu')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
+
         $LoaiSp = LoaiSanPham::find($id);
+
         if(empty($LoaiSp))
         {
             return redirect()->route('loai-san-pham.danh-sach')->with('error','loại sản phẩm không tồn tại');
         }
         $LoaiSp->delete();
+
         return redirect()->route('loai-san-pham.danh-sach')->with('thong_bao','Xóa Thành Công');
- 
      }
  
-     public function CapNhatLoaiSp($id){
+     public function CapNhatLoaiSp($id)
+     {
+        if (Gate::denies('quan-ly-san-pham')) {
+            return redirect()->route('trang-chu')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
          $dsLoaiSp = LoaiSanPham::find($id);
          return view('loai-san-pham/cap-nhat',compact('dsLoaiSp'));
          
@@ -60,5 +75,17 @@ class LoaiSanPhamController extends Controller
          $LoaiSp->save();
          return redirect()->route('loai-san-pham.danh-sach')->with('thong_bao','Cập nhật loại sản phẩm thành công');
      }
+
+     public function Search(Request $request)
+    {
+        if (Gate::denies('quan-ly-san-pham')) {
+            return redirect()->route('trang-chu')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
+        $query = $request->input('query');
+
+        $dsLoaiSp = LoaiSanPham::where('ten_loai', 'LIKE', "%$query%")->get();
+
+        return view('loai-san-pham/danh-sach', compact('dsLoaiSp'));
+    }
 
 }
